Use web3-react getENS
